Moved ViewCount into new package, Circe is here!
